move to pkg/utils
