+++ internal/connect/utils.go
@@ -1,24 +1,13 @@
 package connect
 
 import (
-	"bytes"
 	"encoding/json"
 	"fmt"
-	"io"
-	"io/ioutil"
-	"mime/multipart"
-	"net/http"
 	"os"
-	"path/filepath"
-	"time"
 
 	"github.com/spf13/cobra"
 
 	"github.com/confluentinc/cli/v3/pkg/errors"
-)
-
-const (
-	maxFileSize = 1024 * 1024 * 1024 // 1GB
 )
 
 func getConfig(cmd *cobra.Command) (*map[string]string, error) {
@@ -88,70 +77,4 @@ func parseConfigFile(filename string) (map[string]string, error) {
 	}
 
 	return kvPairs, err
-}
-
-func UploadFile(url, filePath string, formFields map[string]any) error {
-	var buffer bytes.Buffer
-	writer := multipart.NewWriter(&buffer)
-
-	fileInfo, err := os.Stat(filePath)
-	if err != nil {
-		return err
-	}
-
-	fileSize := fileInfo.Size()
-	if fileSize > maxFileSize {
-		return fmt.Errorf("File size exceeds the limit. Maximum allowed size is 1GB. Actual Size %d", fileSize)
-	}
-
-	for key, value := range formFields {
-		if strValue, ok := value.(string); ok {
-			err := writer.WriteField(key, strValue)
-			if err != nil {
-				return err
-			}
-		}
-	}
-
-	file, err := os.Open(filePath)
-	if err != nil {
-		return err
-	}
-	defer file.Close()
-
-	part, err := writer.CreateFormFile("file", filepath.Base(filePath))
-	if err != nil {
-		return err
-	}
-	if _, err := io.Copy(part, file); err != nil {
-		return err
-	}
-
-	if err := writer.Close(); err != nil {
-		return err
-	}
-
-	client := &http.Client{
-		Timeout: 20 * time.Minute,
-	}
-
-	// Create the HTTP request
-	request, err := http.NewRequest("POST", url, &buffer)
-	if err != nil {
-		return err
-	}
-	// Set the Content-Type header to multipart/form-data
-	request.Header.Set("Content-Type", writer.FormDataContentType())
-	response, err := client.Do(request)
-	if err != nil {
-		return err
-	}
-	defer response.Body.Close()
-
-	if response.StatusCode >= 400 {
-		responseBody, err := ioutil.ReadAll(response.Body)
-		return fmt.Errorf("[Response] %s [Error] %v", string(responseBody), err)
-	}
-
-	return nil
 }
